+++ src/main/java/com/ynyes/lyz/entity/TdAd.java
@@ -54,7 +54,7 @@ public class TdAd {
     
     // 排序号
     @Column
-    private Long sortId;
+    private Double sortId;
 
     // 是否使能
     @Column
@@ -128,11 +128,11 @@ public class TdAd {
         this.endTime = endTime;
     }
 
-    public Long getSortId() {
+    public Double getSortId() {
         return sortId;
     }
 
-    public void setSortId(Long sortId) {
+    public void setSortId(Double sortId) {
         this.sortId = sortId;
     }
