+++ src/main/java/com/ynyes/lyz/entity/TdUserLevel.java
@@ -53,7 +53,7 @@ public class TdUserLevel {
     
     // 排序号
     @Column
-    private Long sortId;
+    private Double sortId;
 
     public Long getId() {
         return id;
@@ -103,11 +103,11 @@ public class TdUserLevel {
         this.isEnable = isEnable;
     }
 
-    public Long getSortId() {
+    public Double getSortId() {
         return sortId;
     }
 
-    public void setSortId(Long sortId) {
+    public void setSortId(Double sortId) {
         this.sortId = sortId;
     }
